Start migrating format functions out of the top level module
Push them into more targetted lower level modules
