+++ lib/game/command/look.ex
@@ -10,6 +10,7 @@ defmodule Game.Command.Look do
   alias Game.Environment
   alias Game.Environment.State.Overworld
   alias Game.Environment.State.Room
+  alias Game.Format.Rooms, as: FormatRooms
   alias Game.Hint
   alias Game.Item
   alias Game.Items
@@ -97,7 +98,7 @@ defmodule Game.Command.Look do
          {:ok, room} <- @environment.look(save.room_id),
          %{finish_id: room_id} <- Exit.exit_to(room, direction),
          {:ok, room} <- @environment.look(room_id) do
-      state.socket |> @socket.echo(Format.peak_room(room, direction))
+      state.socket |> @socket.echo(FormatRooms.peak_room(room, direction))
     else
       :overworld ->
         :ok
@@ -137,7 +138,7 @@ defmodule Game.Command.Look do
     state |> GMCP.room(room, items)
     state |> GMCP.map(mini_map)
 
-    state.socket |> @socket.echo(Format.room(room, items, room_map))
+    state.socket |> @socket.echo(FormatRooms.room(room, items, room_map))
 
     maybe_hint_quest(state, room)
   end
@@ -148,7 +149,7 @@ defmodule Game.Command.Look do
     state |> GMCP.map(mini_map)
 
     room = remove_yourself(room, state)
-    state.socket |> @socket.echo(Format.overworld_room(room, mini_map))
+    state.socket |> @socket.echo(FormatRooms.overworld_room(room, mini_map))
   end
 
   defp maybe_hint_quest(state, room) do
