+++ lib/game/format.ex
@@ -7,7 +7,6 @@ defmodule Game.Format do
 
   import Game.Format.Context
 
-  alias Data.Exit
   alias Data.Item
   alias Data.Mail
   alias Data.Room
@@ -15,12 +14,10 @@ defmodule Game.Format do
   alias Data.Save
   alias Data.Skill
   alias Game.Color
-  alias Game.Door
   alias Game.Format.Listen
   alias Game.Format.Resources
   alias Game.Format.Table
   alias Game.Format.Template
-  alias Game.Quest
 
   @doc """
   Template a string
@@ -231,112 +228,10 @@ defmodule Game.Format do
   end
 
   @doc """
-  Format full text for a room
-  """
-  @spec room(Room.t(), [Item.t()], Map.t()) :: String.t()
-  def room(room, items, map) do
-    """
-    #{room_name(room)}
-    #{underline(room.name)}
-    #{room_description(room)}\n
-    #{map}
-
-    #{who_is_here(room)}
-
-    #{maybe_exits(room)}#{maybe_items(room, items)}#{shops(room)}
-    """
-    |> String.trim()
-  end
-
-  @doc """
-  Template a room's description
-  """
-  def room_description(room) do
-    description = room_description_with_features(room)
-
-    context =
-      context()
-      |> assign(:room, "{green}#{room.name}{/green}")
-      |> assign(:zone, "{white}#{room.zone.name}{/white}")
-      |> assign(:features, Enum.join(features(room.features), " "))
-
-    context =
-      Enum.reduce(room.features, context, fn room_feature, context ->
-        assign(context, room_feature.key, feature(room_feature))
-      end)
-
-    template(context, resources(description))
-  end
-
-  defp room_description_with_features(room) do
-    contains_features? = String.contains?(room.description, "[features]")
-    contains_sub_features? = Enum.any?(room.features, fn feature ->
-      String.contains?(room.description, "[#{feature.key}]")
-    end)
-
-    case contains_features? || contains_sub_features? do
-      true ->
-        room.description
-
-      false ->
-        "#{room.description} [features]"
-    end
-  end
-
-  @doc """
-  Display a room's name
-  """
-  def room_name(room) do
-    "{room}#{room.name}{/room}"
-  end
-
-  @doc """
   Display a zone's name
   """
   def zone_name(zone) do
     "{zone}#{zone.name}{/zone}"
-  end
-
-  @doc """
-  Display a room's feature
-  """
-  def feature(feature) do
-    String.replace(feature.short_description, feature.key, "{white}#{feature.key}{/white}")
-  end
-
-  @doc """
-  Display room features
-  """
-  def features(features) do
-    Enum.map(features, &feature/1)
-  end
-
-  @doc """
-  Peak at a room from the room you're in
-
-  Example:
-
-    iex> Game.Format.peak_room(%{name: "Hallway"}, "north")
-    "{room}Hallway{/room} is north."
-  """
-  @spec peak_room(Room.t(), String.t()) :: String.t()
-  def peak_room(room, direction) do
-    "#{room_name(room)} is #{direction}."
-  end
-
-  @doc """
-  Output for an overworld look
-  """
-  @spec overworld_room(Overworld.t(), String.t()) :: String.t()
-  def overworld_room(room, map) do
-    """
-    {bold}#{map}{/bold}
-
-    #{who_is_here(room)}
-
-    #{maybe_exits(room)}
-    """
-    |> String.trim()
   end
 
   @doc """
@@ -408,63 +303,6 @@ defmodule Game.Format do
     Enum.join([str1, str2] |> Enum.reject(&(&1 == "")), joiner)
   end
 
-  defp maybe_exits(room) do
-    case room |> Room.exits() do
-      [] ->
-        ""
-
-      _ ->
-        "Exits: #{exits(room)}\n"
-    end
-  end
-
-  defp exits(room) do
-    room
-    |> Room.exits()
-    |> Enum.sort()
-    |> Enum.map(fn direction ->
-      case Exit.exit_to(room, direction) do
-        %{door_id: door_id, has_door: true} ->
-          "{exit}#{direction}{/exit} (#{Door.get(door_id)})"
-
-        _ ->
-          "{exit}#{direction}{/exit}"
-      end
-    end)
-    |> Enum.join(", ")
-  end
-
-  @doc """
-  Format full text for who is in the room
-
-  Example:
-
-      iex> Game.Format.who_is_here(%{players: [%{name: "Mordred"}], npcs: [%{name: "Arthur", extra: %{status_line: "[name] is here."}}]})
-      "{npc}Arthur{/npc} is here.\\n{player}Mordred{/player} is here."
-  """
-  def who_is_here(room) do
-    [npcs(room), players(room)]
-    |> Enum.reject(fn line -> line == "" end)
-    |> Enum.join("\n")
-  end
-
-  @doc """
-  Format Player text for who is in the room
-
-  Example:
-
-      iex> Game.Format.players(%{players: [%{name: "Mordred"}, %{name: "Arthur"}]})
-      "{player}Mordred{/player} is here.\\n{player}Arthur{/player} is here."
-  """
-  @spec players(Room.t()) :: String.t()
-  def players(%{players: players}) do
-    players
-    |> Enum.map(fn player -> "#{player_name(player)} is here." end)
-    |> Enum.join("\n")
-  end
-
-  def players(_), do: ""
-
   @doc """
   Look at a Player
   """
@@ -474,25 +312,6 @@ defmodule Game.Format do
     |> assign(:name, player_name(player))
     |> template("[name] is here.")
   end
-
-  @doc """
-  Format NPC text for who is in the room
-
-  Example:
-
-      iex> mordred = %{name: "Mordred", extra: %{status_line: "[name] is in the room."}}
-      iex> arthur = %{name: "Arthur", extra: %{status_line: "[name] is here."}}
-      iex> Game.Format.npcs(%{npcs: [mordred, arthur]})
-      "{npc}Mordred{/npc} is in the room.\\n{npc}Arthur{/npc} is here."
-  """
-  @spec npcs(Room.t()) :: String.t()
-  def npcs(%{npcs: npcs}) do
-    npcs
-    |> Enum.map(&npc_status/1)
-    |> Enum.join("\n")
-  end
-
-  def npcs(_), do: ""
 
   @doc """
   The status of an NPC
@@ -514,26 +333,6 @@ defmodule Game.Format do
     |> template(resources(npc.extra.description))
   end
 
-  def maybe_items(room, items) do
-    case Enum.empty?(items) and room.currency == 0 do
-      true ->
-        ""
-
-      false ->
-        "Items: #{items(room, items)}\n"
-    end
-  end
-
-  def items(room, items) when is_list(items) do
-    items = items |> Enum.map(&item_name/1)
-
-    (items ++ [currency(room)])
-    |> Enum.reject(&(&1 == ""))
-    |> Enum.join(", ")
-  end
-
-  def items(_, _), do: ""
-
   @doc """
   Format currency
   """
@@ -543,38 +342,9 @@ defmodule Game.Format do
   def currency(currency) when is_integer(currency), do: "{cyan}#{currency} #{@currency}{/cyan}"
 
   @doc """
-  Format Shop text for shops in the room
-
-  Example:
-
-      iex> Game.Format.shops(%{shops: [%{name: "Hole in the Wall"}]})
-      "Shops: {shop}Hole in the Wall{/shop}\\n"
-
-      iex> Game.Format.shops(%{shops: [%{name: "Hole in the Wall"}]}, label: false)
-      "  - {shop}Hole in the Wall{/shop}"
+  Format items for sale in a shop
   """
-  @spec shops(Room.t()) :: String.t()
-  def shops(room, opts \\ [])
-  def shops(%{shops: []}, _opts), do: ""
-
-  def shops(%{shops: shops}, label: false) do
-    shops
-    |> Enum.map(fn shop -> "  - #{shop_name(shop)}" end)
-    |> Enum.join(", ")
-  end
-
-  def shops(%{shops: shops}, _) do
-    shops =
-      shops
-      |> Enum.map(&shop_name/1)
-      |> Enum.join(", ")
-
-    "Shops: #{shops}\n"
-  end
-
-  def shops(_, _), do: ""
-
-  def list_shop(shop, items), do: Game.Format.Shop.list(shop, items)
+  def list_shop(shop, items), do: Game.Format.Shops.list(shop, items)
 
   @doc """
   Display an item
@@ -867,16 +637,6 @@ defmodule Game.Format do
   end
 
   @doc """
-  Format a quest name
-
-    iex> Game.Format.quest_name(%{name: "Into the Dungeon"})
-    "{quest}Into the Dungeon{/quest}"
-  """
-  def quest_name(quest) do
-    "{quest}#{quest.name}{/quest}"
-  end
-
-  @doc """
   Format an items name, cyan
 
     iex> Game.Format.item_name(%{name: "Potion"})
@@ -1013,68 +773,6 @@ defmodule Game.Format do
   end
 
   @doc """
-  Format the status of a player's quests
-  """
-  @spec quest_progress([QuestProgress.t()]) :: String.t()
-  def quest_progress(quests) do
-    rows =
-      quests
-      |> Enum.map(fn %{status: status, quest: quest} ->
-        [to_string(quest.id), quest.name, quest.giver.name, status]
-      end)
-
-    Table.format("You have #{length(quests)} active quests.", rows, [5, 30, 20, 10])
-  end
-
-  @doc """
-  Format the status of a player's quest
-  """
-  @spec quest_detail(QuestProgress.t(), Save.t()) :: String.t()
-  def quest_detail(progress, save) do
-    %{quest: quest} = progress
-    steps = quest.quest_steps |> Enum.map(&quest_step(&1, progress, save))
-    header = "#{quest.name} - #{progress.status}"
-
-    """
-    #{header}
-    #{header |> underline()}
-
-    #{quest.description |> wrap()}
-
-    #{steps |> Enum.join("\n")}
-    """
-    |> String.trim()
-    |> resources()
-  end
-
-  defp quest_step(step, progress, save) do
-    case step.type do
-      "item/collect" ->
-        current_step_progress = Quest.current_step_progress(step, progress, save)
-        " - Collect #{item_name(step.item)} - #{current_step_progress}/#{step.count}"
-
-      "item/give" ->
-        current_step_progress = Quest.current_step_progress(step, progress, save)
-
-        " - Give #{item_name(step.item)} to #{npc_name(step.npc)} - #{current_step_progress}/#{
-          step.count
-        }"
-
-      "item/have" ->
-        current_step_progress = Quest.current_step_progress(step, progress, save)
-        " - Have #{item_name(step.item)} - #{current_step_progress}/#{step.count}"
-
-      "npc/kill" ->
-        current_step_progress = Quest.current_step_progress(step, progress, save)
-        " - Kill #{npc_name(step.npc)} - #{current_step_progress}/#{step.count}"
-
-      "room/explore" ->
-        current_step_progress = Quest.current_step_progress(step, progress, save)
-        " - Explore #{room_name(step.room)} - #{current_step_progress}"
-    end
-  end
-
-  @doc """
   List out skills that a trainer will give
   """
   @spec trainable_skills(NPC.t(), [Skill.t()]) :: String.t()
@@ -1088,55 +786,6 @@ defmodule Game.Format do
     rows = [["ID", "Name", "Command", "Cost"] | rows]
 
     Table.format("#{npc_name(trainer)} will train these skills:", rows, [5, 30, 20, 10])
-  end
-
-  @doc """
-  Format a list of socials
-  """
-  def socials(socials) do
-    rows =
-      socials
-      |> Enum.map(fn social ->
-        [social.name, "{command}#{social.command}{/command}"]
-      end)
-
-    rows = [["Name", "Command"] | rows]
-
-    Table.format("List of socials", rows, [20, 20])
-  end
-
-  @doc """
-  View a single social
-  """
-  def social(social) do
-    """
-    #{social.name}
-    #{underline(social.name)}
-    Command: {command}#{social.command}{/command}
-
-    With a target: {say}#{social.with_target}{/say}
-
-    Without a target: {say}#{social.without_target}{/say}
-    """
-  end
-
-  @doc """
-  Format the social without_target text
-  """
-  def social_without_target(social, player) do
-    context()
-    |> assign(:user, player_name(player))
-    |> template("{say}#{social.without_target}{say}")
-  end
-
-  @doc """
-  Format the social with_target text
-  """
-  def social_with_target(social, player, target) do
-    context()
-    |> assign(:user, player_name(player))
-    |> assign(:target, name(target))
-    |> template("{say}#{social.with_target}{say}")
   end
 
   @doc """
@@ -1164,36 +813,6 @@ defmodule Game.Format do
       |> Enum.max()
 
     Table.format("Config", rows, [20, max_size])
-  end
-
-  @doc """
-  Format a list of bugs
-  """
-  @spec list_bugs([Bug.t()]) :: String.t()
-  def list_bugs(bugs) do
-    rows =
-      bugs
-      |> Enum.map(fn bug ->
-        [to_string(bug.id), bug.title, to_string(bug.is_completed)]
-      end)
-
-    rows = [["ID", "Title", "Is Fixed?"] | rows]
-
-    Table.format("Bugs", rows, [10, 30, 10])
-  end
-
-  @doc """
-  Format a list of bugs
-  """
-  @spec show_bug(Bug.t()) :: String.t()
-  def show_bug(bug) do
-    """
-    #{bug.title}
-    #{underline(bug.title)}
-    Fixed: #{bug.is_completed}
-
-    #{bug.body}
-    """
   end
 
   @doc """
